Created function components for modal and modal buttons for styling consistency
