+++ client/src/components/Modal/ModalButton/useStyles.ts
@@ -1,8 +1,13 @@
 import { makeStyles } from '@material-ui/core/styles';
 
-const useStyles = makeStyles(() => ({
+const useStyles = makeStyles((theme) => ({
   form: {
     width: '100%', // Fix IE 11 issue.
+  },
+  submit: {
+    margin: theme.spacing(3, 2, 2),
+    width: 160,
+    height: 56,
   },
   circularProgess: { color: 'white' },
 }));
