+++ client/src/components/Modal/ModalPaper/useStyles.ts
@@ -14,13 +14,11 @@ const useStyles = makeStyles(() => ({
     justifyContent: 'space-between',
     flexDirection: 'column',
     paddingTop: 23,
+    maxWidth: '600px',
   },
   welcome: {
     fontSize: 26,
     paddingBottom: 20,
-    color: '#000000',
-    fontWeight: 700,
-    fontFamily: "'Open Sans'",
   },
 }));
