+++ client/src/pages/SignUp/SignUpForm/useStyles.ts
@@ -1,13 +1,8 @@
 import { makeStyles } from '@material-ui/core/styles';
 
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles(() => ({
   form: {
     width: '100%', // Fix IE 11 issue.
-  },
-  submit: {
-    margin: theme.spacing(3, 2, 2),
-    width: 160,
-    height: 56,
   },
   circularProgess: { color: 'white' },
 }));
